+++ test/resources/checkers/elixir-error.ex
@@ -1,7 +1,0 @@
-# Test an error from Elixir
-
-defmodule AnError do
-  def error_func do
-    puts "Flycheck is great!"
-  end
-end