+++ test/resources/checkers/javascript-style.js
@@ -1,0 +1,5 @@
+/** Tab indentation */
+
+(function() {
+	var foo = ["Hello world"];
+}());