+++ test/run.el
@@ -2,9 +2,11 @@
 ":"; exec ${EMACS:-emacs} -Q --script "$0" -- "${@}" # -*- mode: emacs-lisp; lexical-binding: t; -*-
 ;;; run.el --- Flycheck: Test runner
 
-;; Copyright (C) 2014-2015  Sebastian Wiesner <swiesner@lunaryorn.com>
+;; Copyright (C) 2014-2015 Sebastian Wiesner and Flycheck contributors
 
 ;; Author: Sebastian Wiesner <swiesner@lunaryorn.com>
+;; Maintainer: Sebastian Wiesner <swiesner@lunaryorn.com>
+;; URL: https://www.flycheck.org
 
 ;; This file is not part of GNU Emacs.
 
@@ -94,6 +96,11 @@ Node `(ert)Test Selectors' for information about test selectors."
                           (flycheck-run-check-selector selector)
                           (kill-emacs 1))))))
     (ert-run-tests-batch-and-exit (flycheck-transform-selector selector))))
+
+(defun flycheck-setup-coverage-reporting ()
+  "Setup test coverage reporting using undercover.el."
+  (when (fboundp 'undercover)
+    (undercover "flycheck.el")))
 
 (defun flycheck-runs-this-script-p ()
   "Whether this file is executed as script."